#14 Refactor: scss 파일명 통일 및 파일구조 변경
