+++ src/component/memoir/editor/memoirTitleList.tsx
@@ -1,6 +1,6 @@
-import "./memoir-title-list.scss";
+import "./memoirTitleList.scss";
 import { useState } from "react";
-import type { MainTitleItem } from "../memoir.types.ts";
+import type { MainTitleItem } from "./common/memoir.types.ts";
 import MainMamoirTitle from "./mainMemoirTitle.tsx";
 import EditMainMemoirTitle from "./editMainMemoirTitle.tsx";
